requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
aspose-cells-python==25.1.0
babel==2.16.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
cssselect2==0.7.0
Django==5.1.5
et_xmlfile==2.0.0
fonttools==4.55.3
html5lib==1.1
idna==3.10
jpype1==1.5.2
lxml==5.3.0
numpy==2.2.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.1.0
pyphen==0.17.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.3.0
weasyprint==63.1
webencodings==0.5.1
xhtml2pdf==0.2.16
xlsx2html==0.6.2
zopfli==0.2.3.post1
